backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
celery==5.4.0
pydantic-settings==2.5.2
alembic==1.13.3
python-multipart==0.0.9
pytest==8.3.3
httpx==0.27.2
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
boto3==1.35.50
